% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create a Dependencies JSON File}
\usage{
create(
  dir = getwd(),
  file = "dependencies.json",
  sysreqs = FALSE,
  output = dir,
  platform = "DEB",
  installed = c("base", "recommended"),
  overwrite = TRUE
)
}
\arguments{
\item{dir}{Path to the directory where the files to be scanned are located.}

\item{file}{The name of the file to be save, default is \code{"dependencies.json"}.}

\item{sysreqs}{Logical, should system requirements be written in a separate file? Default same as \code{file} with \code{.txt} extension, i.e. \code{"dependencies.txt"}.}

\item{output}{Path to the directory where JSON file should be written to.}

\item{platform}{The platform supplied to \code{sysreqs()}. Can be \code{NULL} or \code{NA} in which case no system requirements are returned.}

\item{installed}{The \code{priority} argument for \code{installed.packages()} for packages to be excluded.}

\item{overwrite}{Logical, should the \code{file} in the \code{output} directory be overwritten if exists?}
}
\value{
Invisibly returns the list of file names that were created. The side effect is a JSON (and possibly a text for system requirements) file written to the hard drive.
}
\description{
Create a Dependencies JSON File
}
\examples{
dir <- system.file("examples/01-basic", package = "deps")
out <- tempdir()
create(dir, output = out)
cat(readLines(file.path(out, "dependencies.json")), sep = "\n")
unlink(file.path(out, "dependencies.json"))

}
