% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps-package.R
\docType{package}
\name{deps-package}
\alias{deps}
\alias{deps-package}
\title{deps: Dependency Management with roxygen-style Comments}
\description{
Manage your source code dependencies by decorating your existing R code with special, roxygen-style comments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/analythium/deps}
  \item Report bugs at \url{https://github.com/analythium/deps/issues}
}

}
\author{
\strong{Maintainer}: Peter Solymos \email{peter@analythium.io} (\href{https://orcid.org/0000-0001-7337-1740}{ORCID})

Other contributors:
\itemize{
  \item Analythium Solutions Inc. [copyright holder, funder]
}

}
\keyword{internal}
