% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Create a Dependencies JSON File}
\usage{
install(
  dir = getwd(),
  file = "dependencies.json",
  upgrade = "never",
  cleanup = TRUE,
  timeout = 300L,
  ...
)
}
\arguments{
\item{dir}{Path to the directory where the JSON file should be written to.}

\item{file}{The name of the file to be save, default is \code{"dependencies.json"}. If the file is not found in \code{dir}, \code{create()} is called.}

\item{upgrade}{Should package dependencies be upgraded? Argument passed to remotes functions.}

\item{cleanup}{Logical, clean up files created by \code{create()} when \code{file} does not exist.}

\item{timeout}{Integer, timeout for file downloads (default 60 seconds can be short).}

\item{...}{Other argument passed to remotes functions.}
}
\value{
Returns \code{NULL} invisibly. The side effect is the dependencies installed.
}
\description{
Create a Dependencies JSON File
}
\examples{
\dontrun{
dir <- system.file("examples/01-basic", package = "deps")
out <- tempdir()
create(dir, output = out)
install(out)
unlink(file.path(out, "dependencies.json"))
}

}
